Plan out next steps
